+++ src/nespresso/bot/handlers/admin/register.py
@@ -1,9 +1,10 @@
 from aiogram import Dispatcher
 
-from nespresso.bot.handlers.admin.commands import logs
+from nespresso.bot.handlers.admin.commands import logs, messages
 
 
 def RegisterAdminHandlers(dp: Dispatcher) -> None:
     dp.include_routers(
         logs.router,
+        messages.router,
     )